# 8317971: RISC-V: implement copySignF/D and signumF/D intrinsics