--- a/src/hotspot/cpu/riscv/c2_MacroAssembler_riscv.cpp
+++ b/src/hotspot/cpu/riscv/c2_MacroAssembler_riscv.cpp
@@ -1661,10 +1661,13 @@ void C2_MacroAssembler::round_double_mode(FloatRegister dst, FloatRegister src, 
 void C2_MacroAssembler::signum_fp(FloatRegister dst, FloatRegister src, FloatRegister one, bool is_double) {
   Register tmp1 = t0;
 
-  Label done, special_val;
+  Label done;
 
   is_double ? fclass_d(tmp1, src)
             : fclass_s(tmp1, src);
+
+  is_double ? fmv_d(dst, src)
+            : fmv_s(dst, src);
 
   //bitmask 0b1100011000 specifies this bits:
   // 3 - src is -0
@@ -1673,16 +1676,11 @@ void C2_MacroAssembler::signum_fp(FloatRegister dst, FloatRegister src, FloatReg
   // 9 - src is a quiet NaN
   andi(tmp1, tmp1, 0b1100011000);
 
-  bnez(tmp1, special_val);
+  bnez(tmp1, done);
 
   // use floating-point 1.0 with a sign of input
   is_double ? fsgnj_d(dst, one, src)
             : fsgnj_s(dst, one, src);
-  j(done);
-
-  bind(special_val);
-  is_double ? fmv_d(dst, src)
-            : fmv_s(dst, src);
 
   bind(done);
 }