--- a/src/hotspot/cpu/riscv/riscv.ad
+++ b/src/hotspot/cpu/riscv/riscv.ad
@@ -7508,9 +7508,8 @@ instruct roundD_reg(fRegD dst, fRegD src, immI rmode, iRegLNoSp tmp1, iRegLNoSp 
 
 // Copysign and signum intrinsics
 
-instruct copySignD_reg(fRegD dst, fRegD src1, fRegD src2, immD0 zero) %{
+instruct copySignD_reg(fRegD dst, fRegD src1, fRegD src2, immD zero) %{
   match(Set dst (CopySignD src1 (Binary src2 zero)));
-  effect(TEMP_DEF dst, USE src1, USE src2);
   format %{ "CopySignD  $dst $src1 $src2" %}
   ins_encode %{
     FloatRegister dst = as_FloatRegister($dst$$reg),
@@ -7523,7 +7522,6 @@ instruct copySignD_reg(fRegD dst, fRegD src1, fRegD src2, immD0 zero) %{
 
 instruct copySignF_reg(fRegF dst, fRegF src1, fRegF src2) %{
   match(Set dst (CopySignF src1 src2));
-  effect(TEMP_DEF dst, USE src1, USE src2);
   format %{ "CopySignF  $dst $src1 $src2" %}
   ins_encode %{
     FloatRegister dst = as_FloatRegister($dst$$reg),
@@ -7534,24 +7532,24 @@ instruct copySignF_reg(fRegF dst, fRegF src1, fRegF src2) %{
   ins_pipe(fp_uop_d);
 %}
 
-instruct signumD_reg(fRegD dst, fRegD src, fRegD zero, fRegD one, iRegINoSp tmp1) %{
+instruct signumD_reg(fRegD dst, fRegD src, immD zero, fRegD one) %{
   match(Set dst (SignumD src (Binary zero one)));
-  effect(TEMP_DEF dst, USE src, USE one, TEMP tmp1);
+  effect(TEMP_DEF dst, USE src, USE one);
   format %{ "signumD  $dst, $src" %}
   ins_encode %{
     __ signum_fp(as_FloatRegister($dst$$reg), as_FloatRegister($src$$reg),
-                 as_FloatRegister($one$$reg), $tmp1$$Register, true /* is_double */);
+                 as_FloatRegister($one$$reg), true /* is_double */);
   %}
   ins_pipe(pipe_class_default);
 %}
 
-instruct signumF_reg(fRegF dst, fRegF src, fRegF zero, fRegF one, iRegINoSp tmp1) %{
+instruct signumF_reg(fRegF dst, fRegF src, immF zero, fRegF one) %{
   match(Set dst (SignumF src (Binary zero one)));
-  effect(TEMP_DEF dst, USE src, USE one, TEMP tmp1);
+  effect(TEMP_DEF dst, USE src, USE one);
   format %{ "signumF  $dst, $src" %}
   ins_encode %{
     __ signum_fp(as_FloatRegister($dst$$reg), as_FloatRegister($src$$reg),
-                 as_FloatRegister($one$$reg), $tmp1$$Register, false /* is_double */);
+                 as_FloatRegister($one$$reg), false /* is_double */);
   %}
   ins_pipe(pipe_class_default);
 %}