Hi all, please review this changes into risc-v floating point copysign and signum intrinsics.
CopySign - returns first argument with the sign of second. On risc-v we have `fsgnj.x` instruction, which can implement this intrinsic.
Signum - returns input value if it is +/- 0.0 or NaN, otherwise 1.0 with the sign of input value returned.  On risc-v we can use `fclass.x` to specify type of input value and return appropriate value.

Tests:
Performance tests on t-head board:
With intrinsics:
```
Benchmark                 (seed)   Mode  Cnt      Score      Error   Units
MathBench.copySignDouble       0  thrpt    8  34156.580 ±   76.272  ops/ms
MathBench.copySignFloat        0  thrpt    8  34181.731 ±   38.182  ops/ms
MathBench.signumDouble         0  thrpt    8  31977.258 ± 1122.327  ops/ms
MathBench.signumFloat          0  thrpt    8  31836.852 ±   56.013  ops/ms
```
Intrinsics turned off (`-XX:+UnlockDiagnosticVMOptions -XX:-UseCopySignIntrinsic -XX:-UseSignumIntrinsic`):
```
Benchmark                 (seed)   Mode  Cnt      Score      Error   Units
MathBench.copySignDouble       0  thrpt    8  31000.996 ±  943.094  ops/ms
MathBench.copySignFloat        0  thrpt    8  30678.016 ±   28.087  ops/ms
MathBench.signumDouble         0  thrpt    8  25435.010 ± 2047.085  ops/ms
MathBench.signumFloat          0  thrpt    8  25257.058 ±   79.175  ops/ms
```
Regression tests: tier1, hotspot:tier2 on risc-v board.

Also, changed name of one micro test: before we had: `sigNumDouble` and `signumFloat` tests, they does not matches to `signum` or `sigNum`. Now we have similar part: `signum`.
Performance tests has been changed a bit, to check intrinsics result better, diff to modify tests:
```
diff --git a/test/micro/org/openjdk/bench/java/lang/MathBench.java b/test/micro/org/openjdk/bench/java/lang/MathBench.java
index 6cd1353907e..0bee25366bf 100644
--- a/test/micro/org/openjdk/bench/java/lang/MathBench.java
+++ b/test/micro/org/openjdk/bench/java/lang/MathBench.java
@@ -143,12 +143,12 @@ public double  ceilDouble() {
 
     @Benchmark
     public double  copySignDouble() {
-        return  Math.copySign(double81, doubleNegative12);
+        return  Math.copySign(double81, doubleNegative12) + Math.copySign(double81, double2) + Math.copySign(double4Dot1, doubleNegative12);
     }
 
     @Benchmark
     public float  copySignFloat() {
-        return  Math.copySign(floatNegative99, float1);
+        return  Math.copySign(floatNegative99, float1) + Math.copySign(eFloat, float1) + Math.copySign(eFloat, floatNegative99);
     }
 
     @Benchmark
@@ -472,12 +472,12 @@ public float  scalbFloatInt() {
 
     @Benchmark
     public double  sigNumDouble() {
-        return  Math.signum(double4Dot1);
+        return  Math.signum(double4Dot1) + Math.signum(doubleNegative12) + Math.signum(double81);
     }
 
     @Benchmark
     public double  signumFloat() {
-        return  Math.signum(floatNegative99);
+        return  Math.signum(floatNegative99) + Math.signum(float2) + Math.signum(float7);
     }
 
     @Benchmark
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8317971](https://bugs.openjdk.org/browse/JDK-8317971): RISC-V: implement copySignF/D and signumF/D intrinsics (**Enhancement** - P4)


### Reviewers
 * [Fei Yang](https://openjdk.org/census#fyang) (@RealFYang - **Reviewer**) ⚠️ Review applies to [867d6e8e](https://git.openjdk.org/jdk/pull/16186/files/867d6e8ea2a53ae71fb8fd341892e8f88347ec07)
 * [Vladimir Kempik](https://openjdk.org/census#vkempik) (@VladimirKempik - Committer) ⚠️ Review applies to [867d6e8e](https://git.openjdk.org/jdk/pull/16186/files/867d6e8ea2a53ae71fb8fd341892e8f88347ec07)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/16186/head:pull/16186` \
`$ git checkout pull/16186`

Update a local copy of the PR: \
`$ git checkout pull/16186` \
`$ git pull https://git.openjdk.org/jdk.git pull/16186/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 16186`

View PR using the GUI difftool: \
`$ git pr show -t 16186`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/16186.diff">https://git.openjdk.org/jdk/pull/16186.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/16186#issuecomment-1761729988)