@Ilyagavrilin 2023-10-19
Change branch inside `signum`, now `dst` always moves to `src`, so we can remove `j(done); `
Performance results:
Without intrinsic:
```
Benchmark               (seed)   Mode  Cnt      Score    Error   Units
MathBench.signumDouble       0  thrpt    8  35666.674 ±  6.317  ops/ms
MathBench.signumFloat        0  thrpt    8  34040.220 ± 13.783  ops/ms
```
With old version:
```
Benchmark               (seed)   Mode  Cnt      Score    Error   Units
MathBench.signumDouble       0  thrpt    8  41601.513 ± 16.570  ops/ms
MathBench.signumFloat        0  thrpt    8  39414.511 ± 28.290  ops/ms
```
With new version:
```
Benchmark               (seed)   Mode  Cnt      Score    Error   Units
MathBench.signumDouble       0  thrpt    8  44060.456 ± 12.483  ops/ms
MathBench.signumFloat        0  thrpt    8  40481.776 ± 28.512  ops/ms
```